test: introduce shared runtime test helpers
