+++ durable-workflow-runtime/src/test/java/com/amannmalik/workflow/runtime/EnterpriseWorkflowE2ETest.java
@@ -13,24 +13,20 @@ import com.amannmalik.workflow.runtime.task.run.RunTaskService;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import dev.restate.common.Output;
 import dev.restate.common.Request;
 import dev.restate.common.Target;
-import dev.restate.sdk.Awakeable;
-import dev.restate.sdk.AwakeableHandle;
+import dev.restate.sdk.WorkflowContext;
+import dev.restate.sdk.common.StateKey;
+import dev.restate.sdk.common.HandlerRequest;
+import dev.restate.serde.TypeTag;
 import dev.restate.sdk.CallDurableFuture;
 import dev.restate.sdk.DurableFuture;
-import dev.restate.sdk.DurablePromise;
-import dev.restate.sdk.DurablePromiseHandle;
 import dev.restate.sdk.InvocationHandle;
-import dev.restate.sdk.RestateRandom;
-import dev.restate.sdk.WorkflowContext;
-import dev.restate.sdk.common.HandlerRequest;
-import dev.restate.sdk.common.RetryPolicy;
-import dev.restate.sdk.common.StateKey;
 import dev.restate.sdk.endpoint.definition.AsyncResult;
 import dev.restate.sdk.endpoint.definition.HandlerContext;
-import dev.restate.serde.TypeTag;
+import dev.restate.sdk.common.RetryPolicy;
+import com.amannmalik.workflow.runtime.testutil.FakeWorkflowContext;
+import com.amannmalik.workflow.runtime.testutil.FakeWorkflowContext;
 import io.serverlessworkflow.api.types.CallTask;
 import io.serverlessworkflow.api.types.EmitTask;
 import io.serverlessworkflow.api.types.ForkTask;
@@ -51,14 +47,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.any;
@@ -91,77 +80,7 @@ class EnterpriseWorkflowE2ETest {
         assertNotNull(ctx);
     }
 
-    static class SimpleAsyncResult<T> implements dev.restate.sdk.endpoint.definition.AsyncResult<T> {
-        private final T value;
-
-        SimpleAsyncResult(T v) {
-            this.value = v;
-        }
-
-        @Override
-        public CompletableFuture<T> poll() {
-            return CompletableFuture.completedFuture(value);
-        }
-
-        @Override
-        public dev.restate.sdk.endpoint.definition.HandlerContext ctx() {
-            return null;
-        }
-
-        @Override
-        public <U> dev.restate.sdk.endpoint.definition.AsyncResult<U> map(dev.restate.common.function.ThrowingFunction<T, CompletableFuture<U>> f, dev.restate.common.function.ThrowingFunction<dev.restate.sdk.common.TerminalException, CompletableFuture<U>> g) {
-            return new SimpleAsyncResult<>(null);
-        }
-    }
-
-    static class SimpleDurableFuture<T> extends DurableFuture<T> {
-        private final T value;
-
-        SimpleDurableFuture(T v) {
-            this.value = v;
-        }
-
-        @Override
-        protected dev.restate.sdk.endpoint.definition.AsyncResult<T> asyncResult() {
-            return new SimpleAsyncResult<>(value);
-        }
-
-        @Override
-        protected Executor serviceExecutor() {
-            return Runnable::run;
-        }
-    }
-
-    static class SimpleInvocationHandle<T> implements InvocationHandle<T> {
-        private final String id;
-
-        SimpleInvocationHandle(String id) {
-            this.id = id;
-        }
-
-        @Override
-        public String invocationId() {
-            return id;
-        }
-
-        @Override
-        public void cancel() {
-        }
-
-        @Override
-        public DurableFuture<T> attach() {
-            return new SimpleDurableFuture<>(null);
-        }
-
-        @Override
-        public Output<T> getOutput() {
-            return Output.ready(null);
-        }
-    }
-
-    static class FakeContext implements WorkflowContext {
-        final Map<String, Object> state = new HashMap<>();
-        final List<Duration> sleeps = new ArrayList<>();
+    static class FakeContext extends FakeWorkflowContext {
         int counter = 0;
 
         @Override
@@ -214,86 +133,26 @@ class EnterpriseWorkflowE2ETest {
 
         @Override
         public <T, R> InvocationHandle<R> send(Request<T, R> request, Duration delay) {
-            return new SimpleInvocationHandle<>("inv-" + (counter++));
+            return new com.amannmalik.workflow.runtime.testutil.SimpleInvocationHandle<>("inv-" + (counter++));
         }
 
         @Override
         public <R> InvocationHandle<R> invocationHandle(String id, TypeTag<R> typeTag) {
-            return new SimpleInvocationHandle<>(id);
+            return new com.amannmalik.workflow.runtime.testutil.SimpleInvocationHandle<>(id);
         }
 
         @Override
         public DurableFuture<Void> timer(String id, Duration duration) {
-            return new SimpleDurableFuture<>(null);
+            return new com.amannmalik.workflow.runtime.testutil.SimpleDurableFuture<>(null);
         }
 
         @Override
         public <T> DurableFuture<T> runAsync(String name, TypeTag<T> typeTag, RetryPolicy policy, dev.restate.common.function.ThrowingSupplier<T> supplier) {
             try {
-                return new SimpleDurableFuture<>(supplier.get());
+                return new com.amannmalik.workflow.runtime.testutil.SimpleDurableFuture<>(supplier.get());
             } catch (Throwable e) {
                 throw new RuntimeException(e);
             }
-        }
-
-        @Override
-        public <T> Awakeable<T> awakeable(TypeTag<T> typeTag) {
-            return null;
-        }
-
-        @Override
-        public AwakeableHandle awakeableHandle(String id) {
-            return null;
-        }
-
-        @Override
-        public RestateRandom random() {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <T> DurablePromise<T> promise(dev.restate.sdk.common.DurablePromiseKey<T> key) {
-            return null;
-        }
-
-        @Override
-        public <T> DurablePromiseHandle<T> promiseHandle(dev.restate.sdk.common.DurablePromiseKey<T> key) {
-            return null;
-        }
-
-        @Override
-        public String key() {
-            return "key";
-        }
-
-        @Override
-        public <T> Optional<T> get(StateKey<T> key) {
-            return Optional.ofNullable((T) state.get(key.name()));
-        }
-
-        @Override
-        public Collection<String> stateKeys() {
-            return state.keySet();
-        }
-
-        @Override
-        public void clear(StateKey<?> key) {
-            state.remove(key.name());
-        }
-
-        @Override
-        public void clearAll() {
-            state.clear();
-        }
-
-        @Override
-        public <T> void set(StateKey<T> key, T value) {
-            state.put(key.name(), value);
-        }
-
-        @Override
-        public void sleep(Duration d) {
-            sleeps.add(d);
         }
 
         private <R> CallDurableFuture<R> completed() {
@@ -306,8 +165,8 @@ class EnterpriseWorkflowE2ETest {
                 ctor.setAccessible(true);
                 return (CallDurableFuture<R>) ctor.newInstance(
                         null,
-                        new SimpleAsyncResult<>(null),
-                        new SimpleDurableFuture<>("id")
+                        new com.amannmalik.workflow.runtime.testutil.SimpleAsyncResult<>(null),
+                        new com.amannmalik.workflow.runtime.testutil.SimpleDurableFuture<>("id")
                 );
             } catch (Exception e) {
                 throw new RuntimeException(e);
